+++ tests/unit/models/data/manifest/craft-manifest-expected.json
@@ -4,8 +4,8 @@
   "component-description": "A fully-defined craft-application project.",
   "component-id": {
     "hashes": {
-      "SHA1": "27d3150b433071ec1e2bd5bf04bc6de92b8b12b5",
-      "SHA256": "9d7f74856a64282de8cb743fafdba600f18eef2a6f6049746b7cb842e47a3123"
+      "sha1": "27d3150b433071ec1e2bd5bf04bc6de92b8b12b5",
+      "sha256": "9d7f74856a64282de8cb743fafdba600f18eef2a6f6049746b7cb842e47a3123"
     }
   },
   "architecture": "amd64",
@@ -19,8 +19,8 @@
       "component-description": "",
       "component-id": {
         "hashes": {
-          "SHA1": "af5834abfa1a537fd383d41f1be33cea47c7b6a7",
-          "SHA256": "38cbbf5467682ada956d14168e301a383d96aaa2f1f694cbaa47cee38b47847d"
+          "sha1": "af5834abfa1a537fd383d41f1be33cea47c7b6a7",
+          "sha256": "38cbbf5467682ada956d14168e301a383d96aaa2f1f694cbaa47cee38b47847d"
         }
       },
       "architecture": "TODO",
@@ -38,8 +38,8 @@
       "component-description": "jammy main Packages file",
       "component-id": {
         "hashes": {
-          "SHA1": "370c66437d49460dbc16be011209c4de9977212d",
-          "SHA256": "37cb57f1554cbfa71c5a29ee9ffee18a9a8c1782bb0568e0874b7ff4ce8f9c11"
+          "sha1": "370c66437d49460dbc16be011209c4de9977212d",
+          "sha256": "37cb57f1554cbfa71c5a29ee9ffee18a9a8c1782bb0568e0874b7ff4ce8f9c11"
         }
       },
       "architecture": "TODO",
@@ -57,8 +57,8 @@
       "component-description": "The Go programming language",
       "component-id": {
         "hashes": {
-          "SHA1": "376506001849698af3f9e07a236a47ee8cddded0",
-          "SHA256": "cf7e02ebfdaa898107d2dbf84cf1231cee6c244dd5646580d09cfd6f6cf12577"
+          "sha1": "376506001849698af3f9e07a236a47ee8cddded0",
+          "sha256": "cf7e02ebfdaa898107d2dbf84cf1231cee6c244dd5646580d09cfd6f6cf12577"
         }
       },
       "architecture": "TODO",
